+++ core-cube/src/main/java/org/apache/kylin/cube/model/validation/rule/AggregationGroupRule.java
@@ -191,7 +191,7 @@ public class AggregationGroupRule implements IValidatorRule<CubeDesc> {
         }
     }
 
-    protected int getMaxCombinations(CubeDesc cubeDesc) {
+    protected long getMaxCombinations(CubeDesc cubeDesc) {
         return cubeDesc.getConfig().getCubeAggrGroupMaxCombination();
     }
 }
